analogWriteFrequency uses float for frequency

+++ teensy3/core_pins.h
@@ -986,7 +986,7 @@ void init_pins(void);
 void analogWrite(uint8_t pin, int val);
 void analogWriteRes(uint32_t bits);
 static inline void analogWriteResolution(uint32_t bits) { analogWriteRes(bits); }
-void analogWriteFrequency(uint8_t pin, uint32_t frequency);
+void analogWriteFrequency(uint8_t pin, float frequency);
 void analogWriteDAC0(int val);
 #ifdef __cplusplus
 void attachInterruptVector(IRQ_NUMBER_t irq, void (*function)(void));
